README.md
Task
Our task was to create a simple query language similar to the well-known SQL but using only the SELECT
command. The query then filtered data written in an XML file which was our input. The result was then written
into an output XML file. This program acts as a script. My implementation is mostly build on regular
expressions making it much shorter than other versions.